# swagger.yaml
swagger: '2.0'
info:
  version: '2.0'
...
  title: SIMILARITY DETECTION COMPONENT
  contact:
    name: UPC-GESSI (OPENReq)
    url: 'http://openreq.eu/'
  license:
    name: License
    url: 'https://www.eclipse.org/org/documents/epl-2.0/EPL-2.0.txt'
host: api.openreq.eu
basePath: /
...
schemes:
  - https
paths:
  /upc/similarity-detection/AddRequirements:
    post:
      tags:
        - Similarity without clusters
      summary: Adds requirements to the tf-idf model of an organization.
      description: |-
        <p><i>Asynchronous</i> method.</p><p>Given a list of requirements, the endpoint pre-processes them and adds them to a specified organization’s tf-idf model. If some of the entry requirements were already part of the model, the endpoint will update its information.
        </p>
      operationId: addRequirementsUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: input
          description: OpenReq JSON with requirements
          required: true
          schema:
            $ref: '#/definitions/Requirements'
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
        - name: url
          in: query
          description: The url where the result of the operation will be returned
          required: false
          type: string
          allowEmptyValue: false
          x-example: 'http://localhost:9406/upload/PostResult'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/BatchProcess:
    post:
      tags:
        - Similarity with clusters
      summary: Updates the organization clusters with the input requirements and dependencies.
      description: '<p><i>Asynchronous</i> method.</p><p>Given a set of updates done in the requirements (see next list), updates the clusters accordingly.</p><p><ul><li>New requirements: The input requirements that do not pertain to the organization''s model are considered to be new requirements. The method stores the pre-processing of the new requirements, adds them to the tf-idf model, and puts the new requirements as clusters of one requirement.</li><li>Updated requirements: The input requirements with a title or text different from the one stored in the database are considered updated requirements. The method updates their pre-processing in the database and updates the organization clusters accordingly.</li><li>New dependencies: The input similarity dependencies that do not pertain to the organization''s model are considered to be new dependencies. The method uses the accepted and rejected dependencies to update the organization clusters.</li><li>Removed dependencies: The organization dependencies that do not appear in the input similarity dependencies are considered removed dependencies. The method updates them as rejected and changes the clusters accordingly.</li></ul></p>'
      operationId: batchProcessUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: input
          description: OpenReq JSON with requirements and dependencies
          required: true
          schema:
            $ref: '#/definitions/ProjectWithDependencies'
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
        - name: url
          in: query
          description: The url where the result of the operation will be returned
          required: false
          type: string
          allowEmptyValue: false
          x-example: 'http://localhost:9406/upload/PostResult'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/BuildClusters:
    post:
      tags:
        - Similarity with clusters
      summary: Generates the clusters using the input requirements and dependencies.
      description: '<p><i>Asynchronous</i> method.</p><p>This method computes the clusters using the existing duplicates. These duplicates relations are defined by the dependencies with type equal to <i>similar</i> or <i>duplicates</i> and status equal to <i>accepted</i>. All the requirements that do not have duplicates relationships with other requirements are considered to be in a cluster of just one requirement. All the requirements are pre-processed and stored in the database, together with their corresponding tf-idf model and the clusters information. The user can choose whether to use only the name of the requirements for constructing the tf-idf model, or to use also the text of the requirements. Finally, all the requirements are compared with all the requirements of other clusters of the organization, and those similarities with score greater than the established threshold for each requirement are stored in the database.</p>'
      operationId: buildClustersUsingPOST
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: formData
          name: compare
          description: Use the text field of the requirements to construct the model
          required: false
          type: boolean
        - name: file
          in: formData
          description: OpenReq JSON with requirements and dependencies
          required: true
          type: file
        - in: formData
          name: organization
          description: Organization
          required: true
          type: string
        - in: formData
          name: threshold
          description: Double between 0 and 1 that establishes the minimum similarity score that the proposed dependencies should have
          required: true
          type: number
          format: double
        - in: formData
          name: url
          description: The url where the result of the operation will be returned
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/BuildClustersAndCompute:
    post:
      tags:
        - Similarity with clusters
      summary: 'Generates the clusters using the input requirements and dependencies, and returns the similarity score between them.'
      description: '<p><i>Asynchronous</i> method.</p><p>This method computes the clusters using the existing duplicates. These duplicates relations are defined by the dependencies with type equal to <i>similar</i> or <i>duplicates</i> and type equal to <i>accepted</i>. All the requirements that do not have duplicates relationships with other requirements are considered to be in a cluster of just one requirement. All the requirements are pre-processed and stored in the database, together with their corresponding tf-idf model and the clusters information. The user can choose whether to use only the name of the requirements for constructing the tf-idf model, or to use also the text of the requirements. Finally, all the requirements are compared with all the requirements of other clusters of the organization, and those similarities with score greater than the established threshold for each requirement are stored in the database. It returns for each requirement the highest similarity score for each cluster (only if they are greater than the established threshold). If the number of maximum dependencies to be returned is received as parameter (maxNumber), the method only returns the <i>maxNumber</i> of dependencies with highest score for each requirement. When <i>maxNumber</i> is equal to 0 only the accepted dependencies are returned and when maxNumber is lower than 0 or not specified, all the accepted and proposed dependencies of each requirement are returned.</p>'
      operationId: buildClustersAndComputeUsingPOST
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: formData
          name: compare
          description: Use the text field of the requirements to construct the model
          required: false
          type: boolean
        - name: file
          in: formData
          description: OpenReq JSON with requirements and dependencies
          required: true
          type: file
        - in: formData
          name: maxNumber
          description: Max number of dependencies to return for each requirement
          required: false
          type: integer
          format: int32
        - in: formData
          name: organization
          description: Organization
          required: true
          type: string
        - in: formData
          name: threshold
          description: Double between 0 and 1 that establishes the minimum similarity score that the proposed dependencies should have
          required: true
          type: number
          format: double
        - in: formData
          name: url
          description: The url where the result of the operation will be returned
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/BuildModel:
    post:
      tags:
        - Similarity without clusters
      summary: Builds a tf-idf model with the input requirements.
      description: '<p><i>Asynchronous</i> method.</p><p>Builds a tf-idf model with the requirements specified in the JSON object. The generated model is assigned to the specified organization and stored in an internal database. Each organization can have only one model. If at the time of generating a new model the corresponding organization already has an existing model, it is replaced by the new one. The user can choose whether to use only the name of the requirements for constructing the model, or to use also the text of the requirements.</p>'
      operationId: buildModelUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: compare
          in: query
          description: Use the text field of the requirements to construct the model
          required: false
          type: boolean
          allowEmptyValue: false
          x-example: true
        - in: body
          name: input
          description: OpenReq JSON with requirements
          required: true
          schema:
            $ref: '#/definitions/Requirements'
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
        - name: threshold
          in: query
          description: Double between 0 and 1 that establishes the minimum similarity score that the proposed dependencies should have
          required: true
          type: number
          format: double
          allowEmptyValue: false
          x-example: 0.1
        - name: url
          in: query
          description: The url where the result of the operation will be returned
          required: false
          type: string
          allowEmptyValue: false
          x-example: 'http://localhost:9406/upload/PostResult'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/BuildModelAndCompute:
    post:
      tags:
        - Similarity without clusters
      summary: Builds a tf-idf model with the input requirements and returns the similarity scores among all the possible pairs of requirements.
      description: '<p><i>Asynchronous</i> method.</p><p>Builds a tf-idf model with the requirements specified in the JSON object.The generated model is assigned to the specified organization and stored in an internal database. Each organization can have only one model. If at the time of generating a new model the corresponding organization already has an existing model, it is replaced by the new one. The user can choose whether to use only the name of the requirements for constructing the model, or to use also the text of the requirements. </p><br><p>This method returns an array of dependencies containing the similarities that are greater than the established threshold between all possible pairs of  requirements received as input.</p>'
      operationId: buildModelAndComputeUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: compare
          in: query
          description: Use the text field of the requirements to construct the model
          required: false
          type: boolean
          allowEmptyValue: false
          x-example: true
        - in: body
          name: input
          description: OpenReq JSON with requirements
          required: true
          schema:
            $ref: '#/definitions/Requirements'
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
        - name: threshold
          in: query
          description: Double between 0 and 1 that establishes the minimum similarity score that the added dependencies should have
          required: true
          type: number
          format: double
          allowEmptyValue: false
          x-example: 0.1
        - name: url
          in: query
          description: The url where the result of the operation will be returned
          required: false
          type: string
          allowEmptyValue: false
          x-example: 'http://localhost:9406/upload/PostResult'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/ClearDatabase:
    delete:
      tags:
        - Auxiliary methods
      summary: Clears all data from the database.
      description: '<p><i>Synchronous</i> method.</p><p>Clears all data from the database. If this method is called while a calculation is being carried out, unforeseen results may occur.</p>'
      operationId: clearDatabaseUsingDELETE
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/DeleteOrganizationData:
    delete:
      tags:
        - Auxiliary methods
      summary: Deletes the organization's requirements data.
      description: '<p><i>Synchronous</i> method.</p><p>Deletes all the models data (tf-idf and/or clusters) of the specified organization. If this method is called while a calculation is being carried out with the chosen organization, unforeseen results may occur.</p>'
      operationId: clearOrganizationUsingDELETE
      produces:
        - application/json
      parameters:
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '404':
          description: Not found
        '500':
          description: Internal error
      deprecated: false
...
  /upc/similarity-detection/DeleteRequirements:
    post:
      tags:
        - Similarity without clusters
      summary: Deletes requirements from the tf-idf model of an organization
      description: |-
        <p><i>Asynchronous</i> method.</p><p>Given a list of requirements, the endpoint deletes them from a specified organization’s tf-idf model. 
        </p>
      operationId: deleteRequirementsUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: input
          description: OpenReq JSON with requirements
          required: true
          schema:
            $ref: '#/definitions/Requirements'
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
        - name: url
          in: query
          description: The url where the result of the operation will be returned
          required: false
          type: string
          allowEmptyValue: false
          x-example: 'http://localhost:9406/upload/PostResult'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/GetOrganizationInfo:
    get:
      tags:
        - Auxiliary methods
      summary: Returns the main organization information
      description: '<p><i>Synchronous</i> method.</p>Returns the main information of the specified organization (name, threshold, if it uses the text attribute and if it has clusters). Also the method returns the async operations that have not finished yet and the pending responses to read.<p></p>'
      operationId: getOrganizationInfoUsingGET
      produces:
        - application/json
      parameters:
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '404':
          description: Not found
        '500':
          description: Internal error
      deprecated: false
  /upc/similarity-detection/GetResponse:
    get:
      tags:
        - Auxiliary methods
      summary: Returns the dependencies that are the result of the other operations.
      description: '<p><i>Synchronous</i> method.</p><p>Returns the result JSON of the asynchronous methods of this API (i.e., all the methods except <i>ReqReq</i>, <i>ReqClusters</i>, <i>TreatAcceptedAndRejectedDependencies</i> and the auxiliary methods). The result is a JSON object formed by a status attribute. If the status is 200 the JSON also contains an array of dependencies which are returned in batches of 20,000 . Each time this operation is called, the following 20,000 dependencies of the indicated response will be returned. An empty JSON will be returned when no more dependencies are left. Nevertheless, if the status attribute is not equal to 200, the json contains the exception message. </p>'
      operationId: getResponsePageUsingGET
      produces:
        - application/json
      parameters:
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
        - name: response
          in: query
          description: Response identifier
          required: true
          type: string
          allowEmptyValue: false
          x-example: '12345678_89'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '404':
          description: Not found
        '423':
          description: The computation is not finished yet
        '500':
          description: Internal error
      deprecated: false
...
  /upc/similarity-detection/TreatAcceptedAndRejectedDependencies:
    post:
      tags:
        - Similarity with clusters
      summary: Updates the organization clusters with the input dependencies.
      description: '<p><i>Synchronous</i> method.</p><p>Given a set of <i>accepted</i> and <i>rejected</i> dependencies, updates the clusters and dependencies accordingly. </p>'
      operationId: treatAcceptedAndRejectedDependenciesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: input
          description: OpenReqJson with dependencies
          required: true
          schema:
            $ref: '#/definitions/Dependencies'
        - name: organization
          in: query
          description: Organization
          required: true
          type: string
          allowEmptyValue: false
          x-example: UPC
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResponseEntity'
        '400':
          description: Bad request
        '500':
          description: Internal error
      deprecated: false
definitions:
  Dependencies:
    type: object
    properties:
      dependencies:
        type: array
        items:
          $ref: '#/definitions/Dependency'
    title: Dependencies
...
  ProjectWithDependencies:
    type: object
    properties:
      dependencies:
        type: array
        items:
          $ref: '#/definitions/Dependency'
      requirements:
        type: array
        items:
          $ref: '#/definitions/Requirement'
    title: ProjectWithDependencies
    description: OpenReqJson with requirements and dependencies
...
  Requirement:
    type: object
    properties:
      created_at:
        type: integer
        format: int64
        example: 1354019441000
      id:
        type: string
        example: UPC-98
      modified_at:
        type: integer
        format: int64
        example: 1354019441000
      name:
        type: string
        example: Check swagger version.
      text:
        type: string
        example: The swagger version is deprecated. Please update the service asap.
    title: Requirement
    description: A requirement with id and text
  Requirements:
    type: object
    properties:
      requirements:
        type: array
        items:
          $ref: '#/definitions/Requirement'
    title: Requirements
    description: OpenReqJson with requirements
...